dodao agregate za korisnika u kontekstu pretplata i za film u kontekstu priprema
