--- src/Movies/Core/Domain/ContributorAggregate/Contributor.cs
@@ -1,20 +1,0 @@
-﻿using Ardalis.GuardClauses;
-using Movies.SharedKernel;
-using Movies.SharedKernel.DomainObjects;
-
-namespace Movies.Core.ContributorAggregate;
-
-public class Contributor : EntityBase, IAggregateRoot
-{
-  public string Name { get; private set; }
-
-  public Contributor(string name)
-  {
-    Name = Guard.Against.NullOrEmpty(name, nameof(name));
-  }
-
-  public void UpdateName(string newName)
-  {
-    Name = Guard.Against.NullOrEmpty(newName, nameof(newName));
-  }
-}
