--- src/Movies/Core/Domain/ContributorAggregate/Events/ContributorDeletedEvent.cs
@@ -1,13 +1,0 @@
-﻿using Movies.SharedKernel.DomainEvents;
-
-namespace Movies.Core.ContributorAggregate.Events;
-
-public class ContributorDeletedEvent : DomainEventBase
-{
-  public int ContributorId { get; set; }
-
-  public ContributorDeletedEvent(int contributorId)
-  {
-    ContributorId = contributorId;
-  }
-}
