--- src/Movies/Core/Domain/ContributorAggregate/Specifications/ContributorByIdSpec.cs
@@ -1,12 +1,0 @@
-﻿using Ardalis.Specification;
-
-namespace Movies.Core.ContributorAggregate.Specifications;
-
-public class ContributorByIdSpec : Specification<Contributor>, ISingleResultSpecification
-{
-  public ContributorByIdSpec(int contributorId)
-  {
-    Query
-        .Where(contributor => contributor.Id == contributorId);
-  }
-}
