--- src/Movies/Core/Domain/Interfaces/IDeleteContributorService.cs
@@ -1,8 +1,0 @@
-﻿using Ardalis.Result;
-
-namespace Movies.Core.Interfaces;
-
-public interface IDeleteContributorService
-{
-    public Task<Result> DeleteContributor(int contributorId);
-}
