--- src/Movies/Core/Domain/Interfaces/IEmailSender.cs
@@ -1,8 +1,0 @@
-﻿using System.Threading.Tasks;
-
-namespace Movies.Core.Interfaces;
-
-public interface IEmailSender
-{
-  Task SendEmailAsync(string to, string from, string subject, string body);
-}
