--- src/Movies/Core/Domain/Interfaces/IToDoItemSearchService.cs
@@ -1,10 +1,0 @@
-﻿using Ardalis.Result;
-using Movies.Core.ProjectAggregate;
-
-namespace Movies.Core.Interfaces;
-
-public interface IToDoItemSearchService
-{
-  Task<Result<ToDoItem>> GetNextIncompleteItemAsync(int projectId);
-  Task<Result<List<ToDoItem>>> GetAllIncompleteItemsAsync(int projectId, string searchString);
-}
