--- src/Movies/Core/Domain/ProjectAggregate/Events/NewItemAddedEvent.cs
@@ -1,16 +1,0 @@
-﻿using Movies.SharedKernel.DomainEvents;
-
-namespace Movies.Core.ProjectAggregate.Events;
-
-public class NewItemAddedEvent : DomainEventBase
-{
-  public ToDoItem NewItem { get; set; }
-  public Project Project { get; set; }
-
-  public NewItemAddedEvent(Project project,
-      ToDoItem newItem)
-  {
-    Project = project;
-    NewItem = newItem;
-  }
-}
