--- src/Movies/Core/Domain/ProjectAggregate/Events/ToDoItemCompletedEvent.cs
@@ -1,13 +1,0 @@
-﻿using Movies.SharedKernel.DomainEvents;
-
-namespace Movies.Core.ProjectAggregate.Events;
-
-public class ToDoItemCompletedEvent : DomainEventBase
-{
-  public ToDoItem CompletedItem { get; set; }
-
-  public ToDoItemCompletedEvent(ToDoItem completedItem)
-  {
-    CompletedItem = completedItem;
-  }
-}
