--- src/Movies/Core/Domain/ProjectAggregate/ProjectStatus.cs
@@ -1,7 +1,0 @@
-﻿namespace Movies.Core.ProjectAggregate;
-
-public enum ProjectStatus
-{
-  InProgress,
-  Complete
-}
