--- src/Movies/Core/Domain/ProjectAggregate/Specifications/IncompleteItemsSearchSpec.cs
@@ -1,14 +1,0 @@
-﻿using Ardalis.Specification;
-
-namespace Movies.Core.ProjectAggregate.Specifications;
-
-public class IncompleteItemsSearchSpec : Specification<ToDoItem>
-{
-  public IncompleteItemsSearchSpec(string searchString)
-  {
-    Query
-        .Where(item => !item.IsDone &&
-        (item.Title.Contains(searchString) ||
-        item.Description.Contains(searchString)));
-  }
-}
