--- src/Movies/Core/Domain/ProjectAggregate/Specifications/IncompleteItemsSpec.cs
@@ -1,11 +1,0 @@
-﻿using Ardalis.Specification;
-
-namespace Movies.Core.ProjectAggregate.Specifications;
-
-public class IncompleteItemsSpec : Specification<ToDoItem>
-{
-  public IncompleteItemsSpec()
-  {
-    Query.Where(item => !item.IsDone);
-  }
-}
