--- src/Movies/Core/Domain/ProjectAggregate/Specifications/ProjectByIdWithItemsSpec.cs
@@ -1,14 +1,0 @@
-﻿using Ardalis.Specification;
-using Movies.Core.ProjectAggregate;
-
-namespace Movies.Core.ProjectAggregate.Specifications;
-
-public class ProjectByIdWithItemsSpec : Specification<Project>, ISingleResultSpecification
-{
-  public ProjectByIdWithItemsSpec(int projectId)
-  {
-    Query
-        .Where(project => project.Id == projectId)
-        .Include(project => project.Items);
-  }
-}
