--- src/Movies/Core/Domain/ProjectAggregate/Specifications/ProjectsWithItemsByContributorId.cs
@@ -1,13 +1,0 @@
-﻿using Ardalis.Specification;
-
-namespace Movies.Core.ProjectAggregate.Specifications;
-
-public class ProjectsWithItemsByContributorIdSpec : Specification<Project>, ISingleResultSpecification
-{
-  public ProjectsWithItemsByContributorIdSpec(int contributorId)
-  {
-    Query
-        .Where(project => project.Items.Where(item => item.ContributorId == contributorId).Any())
-        .Include(project => project.Items);
-  }
-}
