--- src/Movies/Core/Domain/Services/DeleteContributorService.cs
@@ -1,31 +1,0 @@
-﻿using Ardalis.Result;
-using Movies.Core.ContributorAggregate;
-using Movies.Core.ContributorAggregate.Events;
-using Movies.Core.Interfaces;
-using Movies.SharedKernel.Repositories;
-using MediatR;
-
-namespace Movies.Core.Services;
-
-public class DeleteContributorService : IDeleteContributorService
-{
-  private readonly IRepository<Contributor> _repository;
-  private readonly IMediator _mediator;
-
-  public DeleteContributorService(IRepository<Contributor> repository, IMediator mediator)
-  {
-    _repository = repository;
-    _mediator = mediator;
-  }
-
-  public async Task<Result> DeleteContributor(int contributorId)
-  {
-    var aggregateToDelete = await _repository.GetByIdAsync(contributorId);
-    if (aggregateToDelete == null) return Result.NotFound();
-
-    await _repository.DeleteAsync(aggregateToDelete);
-    var domainEvent = new ContributorDeletedEvent(contributorId);
-    await _mediator.Publish(domainEvent);
-    return Result.Success();
-  }
-}
