+++ src/Movies/Core/Domain/ToDoContext/ContributorAggregate/Contributor.cs
@@ -1,0 +1,20 @@
+﻿using Ardalis.GuardClauses;
+using Movies.SharedKernel;
+using Movies.SharedKernel.DomainObjects;
+
+namespace Movies.Domain.ToDoContext.ContributorAggregate;
+
+public class Contributor : EntityBase, IAggregateRoot
+{
+  public string Name { get; private set; }
+
+  public Contributor(string name)
+  {
+    Name = Guard.Against.NullOrEmpty(name, nameof(name));
+  }
+
+  public void UpdateName(string newName)
+  {
+    Name = Guard.Against.NullOrEmpty(newName, nameof(newName));
+  }
+}
