+++ src/Movies/Core/Domain/ToDoContext/ContributorAggregate/Specifications/ContributorByIdSpec.cs
@@ -1,0 +1,12 @@
+﻿using Ardalis.Specification;
+
+namespace Movies.Domain.ToDoContext.ContributorAggregate.Specifications;
+
+public class ContributorByIdSpec : Specification<Contributor>, ISingleResultSpecification
+{
+  public ContributorByIdSpec(int contributorId)
+  {
+    Query
+        .Where(contributor => contributor.Id == contributorId);
+  }
+}
