+++ src/Movies/Core/Domain/ToDoContext/Interfaces/IDeleteContributorService.cs
@@ -1,0 +1,8 @@
+﻿using Ardalis.Result;
+
+namespace Movies.Domain.ToDoContext.Interfaces;
+
+public interface IDeleteContributorService
+{
+  public Task<Result> DeleteContributor(int contributorId);
+}
