+++ src/Movies/Core/Domain/ToDoContext/ProjectAggregate/DomainEventHandlers/ContributorDeletedHandler.cs
@@ -1,0 +1,30 @@
+﻿using Movies.SharedKernel.Repositories;
+using Movies.Domain.ToDoContext.ProjectAggregate.Specifications;
+using MediatR;
+using Movies.Domain.ToDoContext.ContributorAggregate.DomainEvents;
+
+namespace Movies.Domain.ToDoContext.ProjectAggregate.DomainEventHandlers;
+
+public class ContributorDeletedHandler : INotificationHandler<ContributorDeletedEvent>
+{
+  private readonly IRepository<Project> _repository;
+
+  public ContributorDeletedHandler(IRepository<Project> repository)
+  {
+    _repository = repository;
+  }
+
+  public async Task Handle(ContributorDeletedEvent domainEvent, CancellationToken cancellationToken)
+  {
+    var projectSpec = new ProjectsWithItemsByContributorIdSpec(domainEvent.ContributorId);
+    var projects = await _repository.ListAsync(projectSpec);
+    foreach (var project in projects)
+    {
+      project.Items
+        .Where(item => item.ContributorId == domainEvent.ContributorId)
+        .ToList()
+        .ForEach(item => item.RemoveContributor());
+      await _repository.UpdateAsync(project);
+    }
+  }
+}
