+++ src/Movies/Core/Domain/ToDoContext/ProjectAggregate/DomainEvents/NewItemAddedEvent.cs
@@ -1,0 +1,16 @@
+﻿using Movies.SharedKernel.DomainEvents;
+
+namespace Movies.Domain.ToDoContext.ProjectAggregate.DomainEvents;
+
+public class NewItemAddedEvent : DomainEventBase
+{
+  public ToDoItem NewItem { get; set; }
+  public Project Project { get; set; }
+
+  public NewItemAddedEvent(Project project,
+      ToDoItem newItem)
+  {
+    Project = project;
+    NewItem = newItem;
+  }
+}
