+++ src/Movies/Core/Domain/ToDoContext/ProjectAggregate/PriorityStatus.cs
@@ -1,0 +1,12 @@
+﻿using Ardalis.SmartEnum;
+
+namespace Movies.Domain.ToDoContext.ProjectAggregate;
+
+public class PriorityStatus : SmartEnum<PriorityStatus>
+{
+  public static readonly PriorityStatus Backlog = new(nameof(Backlog), 0);
+  public static readonly PriorityStatus Critical = new(nameof(Critical), 1);
+
+  protected PriorityStatus(string name, int value) : base(name, value) { }
+}
+
