+++ src/Movies/Core/Domain/ToDoContext/ProjectAggregate/ProjectStatus.cs
@@ -1,0 +1,7 @@
+﻿namespace Movies.Domain.ToDoContext.ProjectAggregate;
+
+public enum ProjectStatus
+{
+  InProgress,
+  Complete
+}
