+++ src/Movies/Core/Domain/ToDoContext/ProjectAggregate/Specifications/IncompleteItemsSearchSpec.cs
@@ -1,0 +1,14 @@
+﻿using Ardalis.Specification;
+
+namespace Movies.Domain.ToDoContext.ProjectAggregate.Specifications;
+
+public class IncompleteItemsSearchSpec : Specification<ToDoItem>
+{
+  public IncompleteItemsSearchSpec(string searchString)
+  {
+    Query
+        .Where(item => !item.IsDone &&
+        (item.Title.Contains(searchString) ||
+        item.Description.Contains(searchString)));
+  }
+}
