+++ src/Movies/Core/Domain/ToDoContext/ProjectAggregate/Specifications/IncompleteItemsSpec.cs
@@ -1,0 +1,11 @@
+﻿using Ardalis.Specification;
+
+namespace Movies.Domain.ToDoContext.ProjectAggregate.Specifications;
+
+public class IncompleteItemsSpec : Specification<ToDoItem>
+{
+  public IncompleteItemsSpec()
+  {
+    Query.Where(item => !item.IsDone);
+  }
+}
