+++ src/Movies/Core/Domain/ToDoContext/ProjectAggregate/Specifications/ProjectByIdWithItemsSpec.cs
@@ -1,0 +1,13 @@
+﻿using Ardalis.Specification;
+
+namespace Movies.Domain.ToDoContext.ProjectAggregate.Specifications;
+
+public class ProjectByIdWithItemsSpec : Specification<Project>, ISingleResultSpecification
+{
+  public ProjectByIdWithItemsSpec(int projectId)
+  {
+    Query
+        .Where(project => project.Id == projectId)
+        .Include(project => project.Items);
+  }
+}
