+++ src/Movies/Core/Domain/ToDoContext/ProjectAggregate/ToDoItem.cs
@@ -1,0 +1,43 @@
+﻿using Ardalis.GuardClauses;
+using Movies.Domain.ToDoContext.ProjectAggregate.DomainEvents;
+using Movies.SharedKernel.DomainObjects;
+
+namespace Movies.Domain.ToDoContext.ProjectAggregate;
+
+public class ToDoItem : EntityBase
+{
+  public string Title { get; set; } = string.Empty;
+  public string Description { get; set; } = string.Empty;
+  public int? ContributorId { get; private set; }
+  public bool IsDone { get; private set; }
+
+  public void MarkComplete()
+  {
+    if (!IsDone)
+    {
+      IsDone = true;
+
+      RegisterDomainEvent(new ToDoItemCompletedEvent(this));
+    }
+  }
+
+  public void AddContributor(int contributorId)
+  {
+    Guard.Against.Null(contributorId, nameof(contributorId));
+    ContributorId = contributorId;
+
+    var contributorAddedToItem = new ContributorAddedToItemEvent(this, contributorId);
+    base.RegisterDomainEvent(contributorAddedToItem);
+  }
+
+  public void RemoveContributor()
+  {
+    ContributorId = null;
+  }
+
+  public override string ToString()
+  {
+    var status = IsDone ? "Done!" : "Not done.";
+    return $"{Id}: Status: {status} - {Title} - {Description}";
+  }
+}
