+++ src/Movies/Core/Domain/ToDoContext/ProjectAgregateSpecifications.cd
