+++ src/Movies/Core/Domain/ToDoContext/Services/DeleteContributorService.cs
@@ -1,0 +1,31 @@
+﻿using Ardalis.Result;
+using Movies.Domain.ToDoContext.ContributorAggregate.DomainEvents;
+using Movies.SharedKernel.Repositories;
+using MediatR;
+using Movies.Domain.ToDoContext.ContributorAggregate;
+using Movies.Domain.ToDoContext.Interfaces;
+
+namespace Movies.Domain.ToDoContext.Services;
+
+public class DeleteContributorService : IDeleteContributorService
+{
+  private readonly IRepository<Contributor> _repository;
+  private readonly IMediator _mediator;
+
+  public DeleteContributorService(IRepository<Contributor> repository, IMediator mediator)
+  {
+    _repository = repository;
+    _mediator = mediator;
+  }
+
+  public async Task<Result> DeleteContributor(int contributorId)
+  {
+    var aggregateToDelete = await _repository.GetByIdAsync(contributorId);
+    if (aggregateToDelete == null) return Result.NotFound();
+
+    await _repository.DeleteAsync(aggregateToDelete);
+    var domainEvent = new ContributorDeletedEvent(contributorId);
+    await _mediator.Publish(domainEvent);
+    return Result.Success();
+  }
+}
